backend/requirements.txt
# CMD Transparency Backend - Requirements
# Modern FastAPI Backend for Government Procurement Transparency Platform

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# File Handling & CORS Support
python-multipart==0.0.6
aiofiles==23.2.1
pillow==10.4.0

# Database (Async SQLAlchemy 2.x)
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9   # Only needed for synchronous PG tasks

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing
pytest==7.4.3
httpx==0.25.2
